Refactoring

+++ ewm/src/main/java/ru/practicum/explorewithme/ewm/category/service/CategoryServiceImpl.java
@@ -1,13 +1,14 @@
 package ru.practicum.explorewithme.ewm.category.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.explorewithme.ewm.category.db.Category;
 import ru.practicum.explorewithme.ewm.category.db.CategoryRepository;
 import ru.practicum.explorewithme.ewm.common.OffsetLimitPageable;
+import ru.practicum.explorewithme.ewm.exception.ConflictException;
 import ru.practicum.explorewithme.ewm.exception.NotFoundException;
 
 import java.util.List;
@@ -23,20 +24,20 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public Category create(Category category) {
-        return categoryRepository.save(category);
+        return save(category);
     }
 
     @Override
     @Transactional
     public Category update(Category category) {
-        Category oldCategory = getById(category.getId());
+        final Category oldCategory = getById(category.getId());
         oldCategory.setName(category.getName());
-        return categoryRepository.save(oldCategory);
+        return save(oldCategory);
     }
 
     @Override
     public List<Category> getAll(Integer from, Integer size) {
-        Pageable pageable = OffsetLimitPageable.of(from, size);
+        final Pageable pageable = OffsetLimitPageable.of(from, size);
         return categoryRepository.findAll(pageable).getContent();
     }
 
@@ -48,10 +49,19 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
-        try {
+        final Category category = getById(id);
+        if (category.getEvents().isEmpty()) {
             categoryRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException ex) {
-            throw new NotFoundException("Category not found", format("Id=%d", id));
+        } else {
+            throw new ConflictException("Category contains events can not be deleted", "");
+        }
+    }
+
+    private Category save(Category category) {
+        try {
+            return categoryRepository.save(category);
+        } catch (DataIntegrityViolationException ex) {
+            throw new ConflictException("The name is already in use", format("Name=%s", category.getName()));
         }
     }
 }
